E2guardian 3.5.0

+++ src/FOptionContainer.hpp
@@ -474,7 +474,7 @@ class FOptionContainer
 
     int findoptionI(const char *option);
     std::string findoptionS(const char *option);
-    bool realitycheck(int l, int minl, int maxl, const char *emessage);
+    bool realitycheck(long int l, long int minl, long int maxl, const char *emessage);
     int inRegExpURLList(String &url, std::deque<RegExp> &list_comp, std::deque<unsigned int> &list_ref, unsigned int list);
 
     char *inURLList(String &url, unsigned int list, bool doblanket = false, bool ip = false, bool ssl = false);
